Add ScalingMesh to position and scale the Mesh correctly whatever the VisCanvas extent
