+++ src/packages/lib.ts
@@ -22,6 +22,7 @@ export { default as PanZoomMesh } from '../h5web/vis-packs/core/shared/PanZoomMe
 export { default as TooltipMesh } from '../h5web/vis-packs/core/shared/TooltipMesh';
 export { default as ColorBar } from '../h5web/vis-packs/core/heatmap/ColorBar';
 export { default as HeatmapMesh } from '../h5web/vis-packs/core/heatmap/Mesh';
+export { default as HeatmapMeshMaterial } from '../h5web/vis-packs/core/heatmap/MeshMaterial';
 export { default as DataCurve } from '../h5web/vis-packs/core/line/DataCurve';
 export type { VisCanvasProps } from '../h5web/vis-packs/core/shared/VisCanvas';
